Added the sgs_777.txt file
The sgs_777.txt is a simple ground service profile. This was used to enable the simple ground service and select the number of stairs truck, catering and some other ground services vehicles to attach to the plane (like reality). This profile requires Simple Ground Service plugin which can be found in X-Plane.org forums.